nip42 support
