+++ src/Pages/Layout.tsx
@@ -6,13 +6,15 @@ import { faBell, faMessage } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 import { RootState } from "State/Store";
-import { init, setPreferences, UserPreferences } from "State/Login";
+import { init, UserPreferences } from "State/Login";
 import { HexKey, RawEvent, TaggedRawEvent } from "Nostr";
 import { RelaySettings } from "Nostr/Connection";
 import { System } from "Nostr/System"
 import ProfileImage from "Element/ProfileImage";
 import useLoginFeed from "Feed/LoginFeed";
 import { totalUnread } from "Pages/MessagesPage";
+import useEventPublisher from "Feed/EventPublisher";
+import { NIP42AuthChallenge, NIP42AuthResponse } from "Nostr/Auth";
 
 export default function Layout() {
     const dispatch = useDispatch();
@@ -24,10 +26,20 @@ export default function Layout() {
     const readNotifications = useSelector<RootState, number>(s => s.login.readNotifications);
     const dms = useSelector<RootState, RawEvent[]>(s => s.login.dms);
     const prefs = useSelector<RootState, UserPreferences>(s => s.login.preferences);
+    const pub = useEventPublisher();
     useLoginFeed();
 
     useEffect(() => {
         if (relays) {
+            const nip42AuthEvent = async (event: NIP42AuthChallenge) => {
+                if(event.challenge && event.relay) {
+                    const signedEvent = await pub.nip42Auth(event.challenge, event.relay);
+                    const response = new NIP42AuthResponse(event.challenge, signedEvent);
+                    window.dispatchEvent(response);
+                }
+            }
+            window.addEventListener("nip42auth", nip42AuthEvent)
+
             for (let [k, v] of Object.entries(relays)) {
                 System.ConnectToRelay(k, v);
             }
@@ -35,6 +47,10 @@ export default function Layout() {
                 if (!relays[k]) {
                     System.DisconnectRelay(k);
                 }
+            }
+
+            return () => {
+                window.removeEventListener("nip42auth", nip42AuthEvent)
             }
         }
     }, [relays]);
